# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SoftData API
  version: 1.0.0
  description: Public datasets API with optional developer accounts and API keys.
servers:
  - url: https://softdata-api.vercel.app
    description: Production
...
paths:
...
  /v1/finance/microfinance-banks:
    get:
      tags: [Finance]
      operationId: listMicrofinanceBanks
      summary: List Nigerian microfinance banks
      description: Returns the 790-record roster-only CBN/NDIC-reconciled Nigerian microfinance-bank dataset in deterministic name and ID order. No pagination, search, filters, authentication, or deferred enrichment fields are provided in v1.
      responses:
        "200":
          description: Microfinance bank records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicrofinanceBankListResponse"
        "405":
          description: Method not allowed
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/microfinance-banks/{bank_id}:
    get:
      tags: [Finance]
      operationId: getMicrofinanceBank
      summary: Get a Nigerian microfinance bank
      description: Returns one roster-only Nigerian microfinance-bank record identified by its stable lowercase public ID.
      parameters:
        - $ref: "#/components/parameters/MicrofinanceBankID"
      responses:
        "200":
          description: Microfinance bank record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicrofinanceBankResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          description: Method not allowed
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/non-interest-financial-institutions:
    get:
      tags: [Finance]
      operationId: listNonInterestFinancialInstitutions
      summary: List Nigerian non-interest financial institutions
      description: Returns the six-record CBN non-interest financial institution snapshot. No pagination, search, or filters are provided in v1.
      responses:
        "200":
          description: Non-interest financial institution records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NonInterestFinancialInstitutionListResponse"
        "405":
          description: Method not allowed
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/non-interest-financial-institutions/{institution_id}:
    get:
      tags: [Finance]
      operationId: getNonInterestFinancialInstitution
      summary: Get a Nigerian non-interest financial institution
      parameters:
        - $ref: "#/components/parameters/NonInterestFinancialInstitutionID"
      responses:
        "200":
          description: Non-interest financial institution record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NonInterestFinancialInstitutionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          description: Method not allowed
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/merchant-banks:
    get:
      tags: [Finance]
      operationId: listMerchantBanks
      summary: List Nigerian merchant banks
      description: Returns the six-record CBN merchant-bank snapshot. No pagination, search, or filters are provided in v1.
      responses:
        "200":
          description: Merchant bank records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantBankListResponse"
        "405":
          description: Method not allowed
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/merchant-banks/{bank_id}:
    get:
      tags: [Finance]
      operationId: getMerchantBank
      summary: Get a Nigerian merchant bank
      parameters:
        - $ref: "#/components/parameters/MerchantBankID"
      responses:
        "200":
          description: Merchant bank record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantBankResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          description: Method not allowed
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/payment-service-banks:
    get:
      tags: [Finance]
      operationId: listPaymentServiceBanks
      summary: List Nigerian payment service banks
      description: Returns the five-record CBN payment service bank snapshot. No pagination, search, or filters are provided in v1.
      responses:
        "200":
          description: Payment service bank records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentServiceBankListResponse"
        "405":
          description: Method not allowed
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/payment-service-banks/{bank_id}:
    get:
      tags: [Finance]
      operationId: getPaymentServiceBank
      summary: Get a Nigerian payment service bank
      parameters:
        - $ref: "#/components/parameters/PaymentServiceBankID"
      responses:
        "200":
          description: Payment service bank record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentServiceBankResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          description: Method not allowed
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/financial-holding-companies:
    get:
      tags: [Finance]
      operationId: listFinancialHoldingCompanies
      summary: List Nigerian financial holding companies
      description: Returns the seven-record CBN financial holding company snapshot. No pagination, search, or filters are provided in v1.
      responses:
        "200":
          description: Financial holding company records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialHoldingCompanyListResponse"
        "405":
          description: Method not allowed
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/financial-holding-companies/{company_id}:
    get:
      tags: [Finance]
      operationId: getFinancialHoldingCompany
      summary: Get a Nigerian financial holding company
      parameters:
        - $ref: "#/components/parameters/FinancialHoldingCompanyID"
      responses:
        "200":
          description: Financial holding company record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialHoldingCompanyResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          description: Method not allowed
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/development-finance-institutions:
    get:
      tags: [Finance]
      operationId: listDevelopmentFinanceInstitutions
      summary: List Nigerian development finance institutions
      description: Returns the eight-record CBN development finance institution snapshot. No pagination, search, or filters are provided in v1.
      responses:
        "200":
          description: Development finance institution records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevelopmentFinanceInstitutionListResponse"
        "405":
          description: Method not allowed
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/development-finance-institutions/{institution_id}:
    get:
      tags: [Finance]
      operationId: getDevelopmentFinanceInstitution
      summary: Get a Nigerian development finance institution
      parameters:
        - $ref: "#/components/parameters/DevelopmentFinanceInstitutionID"
      responses:
        "200":
          description: Development finance institution record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevelopmentFinanceInstitutionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          description: Method not allowed
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/primary-mortgage-institutions:
    get:
      tags: [Finance]
      operationId: listPrimaryMortgageInstitutions
      summary: List Nigerian primary mortgage institutions
      description: Returns the current 31-record CBN primary mortgage institution snapshot. No pagination, search, or filters are provided in v1.
      responses:
        "200":
          description: Primary mortgage institution records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrimaryMortgageInstitutionListResponse"
        "405":
          description: Method not allowed
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/finance/primary-mortgage-institutions/{institution_id}:
    get:
      tags: [Finance]
      operationId: getPrimaryMortgageInstitution
      summary: Get a Nigerian primary mortgage institution
      parameters:
        - $ref: "#/components/parameters/PrimaryMortgageInstitutionID"
      responses:
        "200":
          description: Primary mortgage institution record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrimaryMortgageInstitutionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          description: Method not allowed
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
            request_id:
              type: string
          required: [code, message, request_id]
      required: [success, error]
...
    MicrofinanceBank:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          maxLength: 128
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        name:
          type: string
        cbn_code:
          type: string
          pattern: '^[0-9]{3}$'
          description: Optional three-digit CBN institution code. Leading zeros are significant.
        nip_code:
          type: string
          pattern: '^[0-9]{6}$'
          description: Optional six-digit NIP receiving-institution code. Leading zeros are significant.
        country_code:
          type: string
          enum: [NG]
      required: [id, name, country_code]
    MicrofinanceBankListResponse:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
        data:
          type: array
          minItems: 790
          maxItems: 790
          items:
            $ref: "#/components/schemas/MicrofinanceBank"
      required: [success, data]
    MicrofinanceBankResponse:
      type: object
      additionalProperties: false
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/MicrofinanceBank"
      required: [success, data]
...
    NonInterestFinancialInstitution:
      type: object
      additionalProperties: false
      properties:
        id: {type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)+$'}
        name: {type: string}
        cbn_code: {type: string, pattern: '^[0-9]{3}$'}
        nip_code: {type: string, pattern: '^[0-9]{6}$'}
        website_url: {type: string, format: uri}
        logo_url: {type: string}
        country_code: {type: string, enum: [NG]}
      required: [id, name, country_code]
...
    FinancialHoldingCompany:
      type: object
      additionalProperties: false
      properties:
        id: {type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)+$'}
        name: {type: string}
        website_url: {type: string, format: uri}
        logo_url: {type: string}
        country_code: {type: string, enum: [NG]}
      required: [id, name, country_code]
    PrimaryMortgageInstitution:
      $ref: "#/components/schemas/NonInterestFinancialInstitution"
    NonInterestFinancialInstitutionListResponse:
      type: object
      additionalProperties: false
      properties:
        success: {type: boolean}
        data:
          type: array
          minItems: 6
          maxItems: 6
          items: {$ref: "#/components/schemas/NonInterestFinancialInstitution"}
      required: [success, data]
    NonInterestFinancialInstitutionResponse:
      type: object
      additionalProperties: false
      properties:
        success: {type: boolean}
        data: {$ref: "#/components/schemas/NonInterestFinancialInstitution"}
      required: [success, data]
    MerchantBankListResponse:
      $ref: "#/components/schemas/NonInterestFinancialInstitutionListResponse"
    MerchantBankResponse:
      $ref: "#/components/schemas/NonInterestFinancialInstitutionResponse"
    PaymentServiceBankListResponse:
      $ref: "#/components/schemas/NonInterestFinancialInstitutionListResponse"
    PaymentServiceBankResponse:
      $ref: "#/components/schemas/NonInterestFinancialInstitutionResponse"
    DevelopmentFinanceInstitutionListResponse:
      $ref: "#/components/schemas/NonInterestFinancialInstitutionListResponse"
    DevelopmentFinanceInstitutionResponse:
      $ref: "#/components/schemas/NonInterestFinancialInstitutionResponse"
    PrimaryMortgageInstitutionListResponse:
      type: object
      additionalProperties: false
      properties:
        success: {type: boolean}
        data:
          type: array
          minItems: 31
          maxItems: 31
          items: {$ref: "#/components/schemas/PrimaryMortgageInstitution"}
      required: [success, data]
    PrimaryMortgageInstitutionResponse:
      type: object
      additionalProperties: false
      properties:
        success: {type: boolean}
        data: {$ref: "#/components/schemas/PrimaryMortgageInstitution"}
      required: [success, data]
    FinancialHoldingCompanyListResponse:
      type: object
      additionalProperties: false
      properties:
        success: {type: boolean}
        data:
          type: array
          minItems: 7
          maxItems: 7
          items: {$ref: "#/components/schemas/FinancialHoldingCompany"}
      required: [success, data]
    FinancialHoldingCompanyResponse:
      type: object
      additionalProperties: false
      properties:
        success: {type: boolean}
        data: {$ref: "#/components/schemas/FinancialHoldingCompany"}
      required: [success, data]
...